move metadata up a dir
