+++ packages/replay/src/main.js
@@ -317,11 +317,13 @@ async function doUploadRecording(
     maybeLog(verbose, `Upload failed: can't connect to server ${server}`);
     return null;
   }
+  // validate metadata before uploading so invalid data can block the upload
+  const metadata = recording.metadata ? buildRecordingMetadata(recording.metadata) : null;
   const recordingId = await connectionCreateRecording(recording.id, recording.buildId);
   maybeLog(verbose, `Created remote recording ${recordingId}, uploading...`);
-  if (recording.metadata) {
+  if (metadata) {
     maybeLog(verbose, `Setting recording metadata for ${recordingId}`);
-    await setRecordingMetadata(recordingId, recording.metadata);
+    await setRecordingMetadata(recordingId, metadata);
   }
   addRecordingEvent(dir, "uploadStarted", recording.id, {
     server,
